UserController is divided on UserAdministrationController - access to the list of users and editing roles and statuses only for the administrator, on UserProfileController - editing user profile, on UserReviewsController (blank) - get the list of user reviews and review editing
